filter screen to babysitter list works perfect

+++ feature_babysitter/src/main/java/com/example/feature_babysitter/babysitter/Babysitter.kt
@@ -9,5 +9,5 @@ data class Babysitter(
     val latitude: Double,
     val longitude: Double,
     val distance : Double,
-    val costPerHour: Int,
+    val costPerHour: Double,
 )
